requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
discord-protos==0.0.2
discord-webhook==1.3.1
discord.py-self==2.0.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
protobuf==5.27.3
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.2
yarl==1.9.4
